vlserver: Split out nvlentry_htonl/ntohl
vlentryread and vlentrywrite contain some code for converting an
nvlentry between in-memory and on-disk representation; the htonl/ntohl
equivalent for an entire nvlentry.

Split out this code into a couple of new functions, nvlentry_htonl and
nvlentry_ntohl, for clarity and to make it easier to reuse this code
in future commits.

Note that these functions are actually the same, since htonl/ntohl are
actually the same operation. But keep two separate functions to try to
make it clear what they are doing.

Change-Id: I7420c7fc7c7afebfea6fb28beb94d8a6a9f04bbc