vlserver: Introduce struct vl_cache
Currently, our ubik application cache for the vlserver is split across
a few different globals (cheader, hostaddress, ex_addr, with rd_* and
wr_* variants). Organize these into a single struct vl_cache, to make
it easier to see what's in the application cache, and to make it
easier to make modifications to our cache management in the future.

Change vl_ctx to point to the relevant vl_cache, instead of pointing
to the individual items in the cache (cheader, et al).

Add 'vldbversion' and 'maxnservers' to the cache, as well. Currently
these are globals and almost never change, but conceptually they are
data that is cached from the on-disk database data.

Change-Id: Ibb31a4c04db7ec0445a985c12ca3dbca721b7577