vlserver: Add higher-level vlwrite/vlread ops
The vlserver already has the vlentryread and vlentrywrite functions
for logically reading/writing a vlentry, but all of the code that
deals with the cheader and extent blocks just uses raw vlread/vlwrite.

To abstract the cheader/exaddr code a bit, introduce a few new
functions for logically reading and writing the cheader and exaddr
blocks: vlread_cheader, vlwrite_cheader, vlread_exblock, and
vlwrite_exblock.

The vlread_* variants always just read in the whole item, but the
vlwrite_* functions allow for writing just small portions of the
relevant item to disk. The way this works is that we pass a pointer to
the object itself, in addition to a separate pointer to the portion we
want to write. This allows the vlwrite_* function to calculate the
offset, avoiding the need for each caller to do so.

As a result of this, vlwrite() is now only called inside vlutils.c, so
make it static.

Change-Id: I9e223a8e881c768ae5462467fd040d508c825528