ubik: Refactor db recovery sending/receiving
Currently, our recovery code that implements the client-side and
server-side variants of DISK_SendFile and DISK_GetFile involves a lot
of duplicated logic. This includes the i/o loop for receiving/sending
the database from the net, setting various flags and locks, pivoting a
new db file into place, etc. There are also subtle differences in some
of the duplicated code, which is hard to see.

Having two implementations of basically the same codepath makes it
difficult to make changes to the recovery process. So to eliminate the
duplicated code, and ensure that both variants of the relevant code
paths are mostly the same, split out the common code into two new
high-level functions: urecovery_send_db and urecovery_receive_db.

This commit contains a few other changes to implement this:

- A couple of database functions get _path variants (uphys_stat_path,
  uphys_setlabel_path), to operate on a database file in the given
  path, instead of the in-use ubik_dbase database file. (The old
  uphys_stat is now removed, since it is unused.)

- When receiving a database from the net into a temporary file, we
  setlabel the temporary file before pivoting the file into place,
  instead of afterwards.

- This commit adds a few higher-level functions to deal with database
  files on disk, in a new subsystem called udb (src/ubik/udb.c).

Change-Id: Id103cdf286b4a731ad110994d9b34c6a079a7e6c