Introduce liboafs_okv
Introduce a new internal library "okv" for interacting with key-value
("kv") databases on local disk. Currently, the only backend
implemented is lmdb, so this is effectively a wrapper around lmdb to
provide additional functionality (such as being able to use a write
transaction in different threads, and reference-counted database
handles). But of course other backends may be implemented in the
future (e.g. sqlite, *dbm).

This also defines a small meta-format for "okv" databases. That is, an
okv database located in /tmp/data.db consists of the following:

- A directory, /tmp/data.db

- A configuration file, /tmp/data.db/oafs-storage.conf, an ini-style
  file containing which backend the database uses, as well as optional
  backend-specific or application-specific data

- Backend-specific data inside /tmp/data.db/ (such as data.mdb and
  lock.mdb for lmdb)

This commit implements okv as a pthreads-only library. For
convenience, okv.h defines LWP stubs in the header to return errors
for all LWP callers without needing to link to any additional
libraries.

Change-Id: If0f3d541be02dd7ae92b7bf35357fb1959e2fcc6