Introduce afsctl
Add a local-machine RPC scheme based on unix domain sockets, called
afsctl (or 'ctl' for short). This is similar in purpose to the SYNC
protocol used by the fileserver daemons, but not tied to the volume
package, and without SYNC's various restrictions:

- The server can handle multiple requests in parallel (SYNC just has
  one synchronous server thread)

- Data is encoded in JSON (SYNC uses flat structs)

The goal is to use afsctl as an easy way to interact with any local
daemon via user-facing tools. The afsctl protocol itself is intended
as an internal non-public interface, and so can be changed more easily
without concern for backwards compatibility, capability negotiation,
etc. Performance is not a concern, since the relevant operations
should be administrative operations, and somewhat infrequent (compared
to e.g. Rx RPCs).

Compared to using Rx, some of the main practical differences are:

- No remote network access

- No authentication, encryption, etc

- No authorization beyond controlling who can access the socket
  (typically root only)

- No need to handle cases like timeouts; we know immediately if a peer
  dies

- No impact on server Rx traffic

- No XDR; protocol changes and RPC in/out arguments can be changed
  more easily at will

This commit just adds the base subsytem and a little bit of
integration glue for vlserver and ptserver, but no actual users
besides some tests. Future commits will add actual functionality for
various servers or subsystems.

Change-Id: If420701fe0db12404630db0f7675b31b633b21ef