auth: Introduce afsconf_ParseNetFiles_int
Change all callers of afsconf_ParseNetFiles to call a new function,
afsconf_ParseNetFiles_int, so they don't need to explicitly pass the
paths to the NetInfo and NetRestrict files.

Also remove the logic in some callers to check if
AFSDIR_SERVER_NETINFO_FILEPATH or AFSDIR_SERVER_NETRESTRICT_FILEPATH
are non-NULL, and if we're on non-NT40. These paths are always not
NULL, and this functionality exists on all platforms; some code paths
already call this function unconditionally.

Create the new header file cellconfig_np.h for this new function.

Change-Id: I2c58fc57cd4f91b5feb3962afdb9970eae19626b