Introduce 'vos -vlserver' and 'pts -ptserver'
Introduce an option to the 'vos' command suite ('-vlserver'), which
allows the user to force the use of a specific vlserver to contact.
This is very useful to have when testing environments with multiple
dbservers, but can also be useful to users if they suspect that
specific dbservers are behaving weirdly.

Do the same thing for 'pts', with a corresponding option of
'-ptserver'.

Change-Id: I7fb2f54dce83028af26beb392bd9d2493cde587e