ubik: Use rxBind host as myHost
Currently, ubik servers always use the results of gethostname() to get
the canonical IP of the current host. But if we're using -rxbind to
bind to a single IP, then we're supposed to treat that as the only IP
for the host. So use that IP as the canonical address for the host
(myHost) when we initialize ubik.

Change-Id: Icc3bfb640f884fcb89ecc303f19f5557d506918f