tests: Introduce afstest_ForkRxProc
Currently, the auth/superuser-t tests use a custom Rx service, and
spawn a separate process to run the server. This file contains a bunch
of logic for spawning the separate process, and signaling when the
server is ready to receive requests, handling process exit gracefully,
etc.

Move this logic into an afstest_* function, called afstest_ForkRxProc.
Change the signaling method to use pipes instead of signals to avoid
dealing with signal masks and the hacky waitforsig() sleep/retry loop,
and to get notification when the peer of the pipe unexpectedly dies.

While we're here, also remove some unnecessary gethostbyname() logic.

Change-Id: I611073afecb1ff4821ee5e7a7c840de124075520