rx: Count buffered data in call->app.bytesSent
Currently, when someone calls rx_Write (or rx_Write32, rx_Writev), we
don't count the given data towards the call->app.bytesSent counter
immediately; the bytesSent counter only gets incremented when we
prepare to send a packet on the wire.

Most of the time this doesn't matter, since the data will eventually
get counted in bytesSent as the packet fills up, or the call is
flushed, etc. We don't look at the bytesSent counter until
rx_RecordCallStatistics() is called, which happens near the end of
running an RPC.

However, there are at least a couple of situations where these bytes
never get counted. The most common situation is on an Rx server, where
we always call rx_RecordCallStatistics() before running rx_EndCall().
When rx_RecordCallStatistics is called and we have outgoing data
buffered in the call that hasn't been flushed, that data is not
reflected in the bytesSent counter, and so is never recorded.

This can easily be seen when looking at the process/peer statistics,
since all of the server-side bytesSent counters are 0 for all RPCs
where the output arguments are small (where we send less than a
packet's worth of data back to the client). For RPCs where the server
tends to send more than a packets worth of data, e.g. RXAFS_FetchData
or AFSVolDump, some data will be recorded, but bytesSent will still
usually be undercounted by the last packet's worth of data.

Another situation where bytesSent can miss data is on clients issuing
a split RPC where the server never sends anything to the client. For
these calls, we call rx_RecordCallStatistics before rx_EndCall, and
since the server never sends us any data, we call may have outgoing
data buffered in the call, which is never counted. Split calls with no
server-sent data are somewhat rare, but they do exist. Some examples
in the current OpenAFS tree are: BOZO_Install, DISK_SendFile, and
AFSVolRestore.

For non-split calls on clients, this issue does not happen, since we
call rx_RecordCallStatistics after rx_EndCall (!), and so any buffered
data will be flushed before recording statistics. Split calls that
receive data from the server will be in receiving mode by the time we
call rx_RecordCallStatistics, and so we cannot be buffering any
outgoing data.

To solve this issue, just change call->app.bytesSent to be incremented
as soon as rx_Write returns (or its variants). This is slightly
inaccurate, since we may have not actually sent the data on the wire
yet, but we never actually look at the counter until the call ends, so
this usually doesn't matter. This can slightly change the call
statistics in the case of aborted calls, but of course aborted calls
are usually rare.

There are some other possible ways to address this issue:

- For server-side calls, we could flush the call before recording
  statistics. This is a bit more iffy to do for client-side calls,
  since some client code records stats after ending the call (though
  this is itself questionable).

- We could defer recording call stats to when the call is ending
  (inside rx_EndCall), and merely use functions like
  rx_RecordCallStatistics to store what call to record statistics for.
  But again, this doesn't fix existing client code that records stats
  after ending the call.

Updating bytesSent on rx_Write return handles all of these cases, and
tends to be conceptually simpler, since we don't need to worry so much
about "when" the counter is updated or when we need to flush the call,
etc.

Credit to mvitale@sinenomine.net, who originally pointed out the
missing data for server-side calls.

Change-Id: Id3266d6f10212468f3ea79c4ea8069e301f90a59