GSoC 2022: Per volume reverse name lookup index
Currently, OpenAFS has "File ID" or FID to "entry" or name lookup
functionality, AKA, reverse lookup and also normal lookup, which is “name”
for a given FID. Inverse Lookup is currently computationally expensive because
the algorithm scans a hash index data structure embedded within a
"Directory object" (which has key:name, value: FID) looking for a name that
matches the given FID.
It is not performant as it's merely brute-forcing the search.

This commit has code changes which add a completely new feature, a key-value
database built on top of LMDB, and along with it low-level unit tests.

The key is defined as 'struct ridb_key'.