GSoC 2022: RIDB - Add new RPC for reverse lookup
This commit adds a new RPC* to lookup the reverse index database for "name"
given the FID. Its called 'InverseLookup2'.

Definition:
InverseLookup2(
   IN AFSFid *Fid,
   OUT string Filename<AFSNAMEMAX>,
   AFSFid *ParentFid
 )

It takes the FID as input and returns the parent FID and filename associated
to the given FID.

The RPC* allows the reverse index database to be queried by a client.

*NOTE: This RPC is a non-standard RPC that we're just using for testing the
RIDB (reverse index database) functionality for now.