GSoC 2022: RIDB - Python script for RPC test
This commit adds Andrew's Python script to test the new RPC without making any
changes to the AFS client.
This can be used to perform barebones testing and should not be used as a
robust/clean way to test the entire setup.

The test python script is also a part of keyword "Get Name by FID" in
robotframework-openafslibrary.